Implemented Report Screen, removed report option from drawer

+++ app/src/main/java/mk/vozenred/bustimetableapp/navigation/destinations/RelationsComposable.kt
@@ -13,7 +13,7 @@ fun NavGraphBuilder.relationsComposable(
   sharedViewModel: SharedViewModel,
   navigateToSearchScreen: () -> Unit,
   navigateToContactScreen: () -> Unit,
-  navigateToReportScreen: () -> Unit
+  navigateToReportScreen: (Int) -> Unit
 ) {
   composable(
     route = RELATIONS_SCREEN,
